Initial Pass - Scripts
Used Godot 4.0.rc2's project upgrade tool.

Updated .gitignore; The godot/, .mono/, and .import/ folders are all now stored in .godot/

Resolved all of the explicit syntax errors in the code:
- Following Godot's lead, swapped 32 bit numeric types out for their 64 bit equivalents (this process is still ongoing but the minimum necessary to compile has been done)
- The project upgrade tool favored sticking with the .Connect() method for wiring up signals; Refactored those to the C# Action pattern.
- Refactored the InitAndConnect method to account for the switch to the C# Action pattern for signal handlers; Created a new Init* method for each individual type as method overloading wasn't resolving correctly for every Node.
- Updated OCSM.Transport<T> to extend GodotObject; Godot 4 renames Godot.Object to GodotObject, in the C# API at least.
- Added ICanDelete interface to facilitate the functionality of NodeUtilities.displayDeleteConfirmation()
- Added partial keyword to every class that extends GodotObject or Node.
- Cleaned up most of the incorrect automatic renames performed by the project upgrade tool; There may be more but the code at least compiles now.

+++ scripts/data/cod/Mortal.cs
@@ -4,8 +4,8 @@ namespace OCSM.CoD
 {
 	public class Mortal : CodCore
 	{
-		public int Age { get; set; }
-		public int Integrity { get; set; }
+		public long Age { get; set; }
+		public long Integrity { get; set; }
 		public string Faction { get; set; }
 		public string GroupName { get; set; }
 		public string Vice { get; set; }
